refactor: Renderer관련 클래스들 패키지 이동
