+++ src/main/java/chess/controller/BoardDto.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.controller;
 
 import java.util.List;
