+++ src/main/java/chess/view/OutputView.java
@@ -1,5 +1,7 @@
 package chess.view;
 
+import chess.controller.BoardDto;
+
 import java.util.List;
 
 public class OutputView {
